misc: move project conversion into separate file
