+++ crates/pixi-build-backend/src/cli.rs
@@ -3,7 +3,6 @@ use std::path::{Path, PathBuf};
 use clap::{Parser, Subcommand};
 use clap_verbosity_flag::{InfoLevel, Verbosity};
 use miette::{Context, IntoDiagnostic};
-use pixi_build_type_conversions::to_project_model_v1;
 use pixi_build_types::{
     procedures::{
         conda_build::CondaBuildParams,
@@ -12,7 +11,6 @@ use pixi_build_types::{
         negotiate_capabilities::NegotiateCapabilitiesParams,
     },
     BackendCapabilities, ChannelConfiguration, FrontendCapabilities, PlatformAndVirtualPackages,
-    VersionedProjectModel,
 };
 use rattler_build::console_utils::{get_default_env_filter, LoggingOutputHandler};
 use rattler_conda_types::{ChannelConfig, GenericVirtualPackage, Platform};
@@ -22,6 +20,7 @@ use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
 
 use crate::{
     consts,
+    project::to_project_model,
     protocol::{Protocol, ProtocolInstantiator},
     server::Server,
 };
@@ -122,56 +121,6 @@ pub async fn main<T: ProtocolInstantiator, F: FnOnce(LoggingOutputHandler) -> T>
             Ok(())
         }
     }
-}
-
-/// Convert manifest to project model
-fn to_project_model(
-    manifest_path: &Path,
-    channel_config: &ChannelConfig,
-    highest_supported_project_model: Option<u32>,
-) -> miette::Result<Option<VersionedProjectModel>> {
-    // Load the manifest
-    let manifest =
-        pixi_manifest::Manifests::from_workspace_manifest_path(manifest_path.to_path_buf())?;
-    let package = manifest.value.package.as_ref();
-
-    // Determine which project model version to use
-    let version_to_use = match highest_supported_project_model {
-        // If a specific version is requested, use it (or fail if it's higher than what we support)
-        Some(requested_version) => {
-            let our_highest = VersionedProjectModel::highest_version();
-            if requested_version > our_highest {
-                miette::bail!(
-                    "Requested project model version {} is higher than our highest supported version {}",
-                    requested_version,
-                    our_highest
-                );
-            }
-            // Use the requested version
-            requested_version
-        }
-        // If no specific version is requested, use our highest supported version
-        None => VersionedProjectModel::highest_version(),
-    };
-
-    // This can be null in the rattler-build backend
-    let versioned = package
-        .map(|manifest| {
-            let result = match version_to_use {
-                1 => to_project_model_v1(&manifest.value, channel_config)
-                    .expect("failed to convert manifest to project model"),
-                _ => {
-                    miette::bail!(
-                        "Unsupported project model version: {}",
-                        VersionedProjectModel::highest_version()
-                    );
-                }
-            };
-            Ok(VersionedProjectModel::from(result))
-        })
-        .transpose()?;
-
-    Ok(versioned)
 }
 
 /// Negotiate the capabilities of the backend and initialize the backend.
