+++ crates/pixi-build-backend/src/lib.rs
@@ -4,6 +4,7 @@ pub mod server;
 
 mod consts;
 pub mod dependencies;
+pub mod project;
 pub mod source;
 pub mod tools;
 pub mod traits;
