+++ crates/pixi-build-backend/src/protocol.rs
@@ -9,7 +9,6 @@ use pixi_build_types::procedures::{
 /// and endpoint that can handle the RPC calls.
 #[async_trait::async_trait]
 pub trait ProtocolInstantiator: Send + Sync + 'static {
-
     /// Called when negotiating capabilities with the client.
     /// This is determine how the rest of the initialization will proceed.
     async fn negotiate_capabilities(
